requirements.txt
aiodns==3.2.0
aiohttp==3.9.4
aiosignal==1.3.1
annotated-types==0.6.0
anyio==4.3.0
async-timeout==4.0.3
asyncio==3.4.3
attrs==23.2.0
beautifulsoup4==4.12.3
bonbast==1.0.2
Brotli==1.1.0
bs4==0.0.2
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
exceptiongroup==1.2.0
fastapi==0.110.1
frozenlist==1.4.1
h11==0.14.0
httptools==0.6.1
idna==3.7
markdown-it-py==3.0.0
mdurl==0.1.2
multidict==6.0.5
pycares==4.4.0
pycparser==2.22
pydantic==2.7.0
pydantic_core==2.18.1
Pygments==2.17.2
PyJWT==2.8.0
python-dotenv==1.0.1
PyYAML==6.0.1
redis==5.0.3
requests==2.31.0
rich==13.7.1
sniffio==1.3.1
soupsieve==2.5
starlette==0.37.2
typing_extensions==4.11.0
urllib3==2.2.1
uvicorn==0.29.0
uvloop==0.19.0
watchfiles==0.21.0
websockets==12.0
yarl==1.9.4
